Tu20 cannonball dev (#26)
* did a rough frame for _actionTank need colliision and cannon ball inorder to continue

* tu20: created file structure

* tu20: BASIC pseudocode

* comment out some things to work 1 element at a time

* added the shooting for the cannonball

* tu20: preliminary cannonball done

* tu20: cannonball fires

* tu20: code cleanup and new impl

* resolved issue with variable exceeding its storage space

---------

+++ menu.c
@@ -16,7 +16,7 @@ bool isPlaying = false;
 CP_Font font;
 CP_Image menuBg;
 
-BYTE menuState = FADE_IN;
+int menuState = FADE_IN;
 
 BYTE oAlpha = 255;
